+++ app/Models/PostsModel.php
@@ -15,11 +15,11 @@ class PostsModel extends Model
     protected $updatedField  = 'updated_at';
     protected $deletedField  = 'deleted_at';*/
 
-    public function getPosts($slug = false)
-{
-    if ($slug === false)
-    {
-        return $this->findAll();
+    public function getPosts($slug = false, $paginate = null){
+    if ($slug === false && is_null($paginate)){
+       return $this->orderBy('created_at', 'DESC')->findAll();       
+    }else if (!is_null($paginate) && $paginate > 0){
+        return $this->orderBy('created_at', 'DESC')->paginate($paginate,'blog');
     }
     return $this->asArray()
                 ->where(['slug' => $slug])