+++ app/Views/blog/template.php
@@ -80,16 +80,10 @@
 	    </section>
 	    <section class="blog-list px-3 py-5 p-md-5">
 		    <div class="container">
-				<?php
-					if(isset($view)){
-						echo view($view);
-					}
-				?>
-			    <nav class="blog-nav nav nav-justified my-5">
-				  <a class="nav-link-prev nav-item nav-link d-none rounded-left" href="#">Previous<i class="arrow-prev fas fa-long-arrow-alt-left"></i></a>
-				  <a class="nav-link-next nav-item nav-link rounded" href="blog-list.html">Next<i class="arrow-next fas fa-long-arrow-alt-right"></i></a>
-				</nav>
-				
+				<?= $this->renderSection('content') ?>				
+				<div class="d-flex justify-content-center">
+					<?= $pager->links('blog','blog_pagination') ?>
+				</div>				
 		    </div>
 	    </section>